(refactor) criar pasta categorias
